chore: move display-objects
