+++ src/feature/area-chart/display-objects/index.ts
@@ -1,13 +1,3 @@
 export * from "./area";
-export * from "./crosshair";
-export * from "./horizontal-axis";
-export * from "./horizontal-grid";
-export * from "./horizontal-line";
-export * from "./indicator";
-export * from "./label";
 export * from "./line-curve";
 export * from "./price-curve";
-export * from "./rect";
-export * from "./vertical-axis";
-export * from "./vertical-grid";
-export * from "./vertical-line";
