+++ src/feature/area-chart/ui.ts
@@ -12,10 +12,6 @@ import { InteractionEvent } from "../../renderer/interaction/interaction-event";
 import { Rectangle } from "../../renderer/math";
 import { hex2string } from "../../renderer/utils";
 import { ScaleLinear, ScaleTime } from "../../types";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
-import { Gesture, Zoom, zoomIdentity, ZoomTransform } from "../../util/zoom";
-import { Data } from "./area-chart";
-import { SeriesData } from "./chart";
 import {
   Crosshair,
   HorizontalAxis,
@@ -23,7 +19,10 @@ import {
   Label,
   pointer,
   VerticalAxis,
-} from "./display-objects";
+} from "../../ui/display-objects";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
+import { Gesture, Zoom, zoomIdentity, ZoomTransform } from "../../util/zoom";
+import { SeriesData } from "./chart";
 import { Colors } from "./helpers";
 
 type UiColors = Pick<
