+++ src/feature/price-chart/display-objects/index.ts
@@ -1,12 +1,2 @@
 export * from "./area";
-export * from "./crosshair";
-export * from "./horizontal-axis";
-export * from "./horizontal-grid";
-export * from "./horizontal-line";
-export * from "./indicator";
-export * from "./label";
 export * from "./price-curve";
-export * from "./rect";
-export * from "./vertical-axis";
-export * from "./vertical-grid";
-export * from "./vertical-line";
