+++ src/feature/price-chart/ui.ts
@@ -11,8 +11,6 @@ import { InteractionEvent } from "../../renderer/interaction/interaction-event";
 import { Rectangle } from "../../renderer/math";
 import { hex2string } from "../../renderer/utils";
 import { ScaleLinear, ScaleTime } from "../../types";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
-import { Gesture, Zoom, zoomIdentity, ZoomTransform } from "../../util/zoom";
 import {
   Crosshair,
   HorizontalAxis,
@@ -20,7 +18,9 @@ import {
   Label,
   pointer,
   VerticalAxis,
-} from "./display-objects";
+} from "../../ui/display-objects";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
+import { Gesture, Zoom, zoomIdentity, ZoomTransform } from "../../util/zoom";
 import { Colors } from "./helpers";
 import { Data } from "./price-chart";
