+++ src/ui/display-objects/horizontal-axis.ts
@@ -1,14 +1,13 @@
-import { range } from "../../../helpers";
-import { Renderer } from "../../../renderer";
-import { Container } from "../../../renderer/display";
-import { Graphics } from "../../../renderer/graphics";
-import { InteractionEvent } from "../../../renderer/interaction/interaction-event";
-import { Text } from "../../../renderer/text";
-import { ScaleTime } from "../../../types";
-import { Colors } from "../../../components/depth-chart/helpers";
-import { AXIS_HEIGHT, FONT_SIZE } from "../../../util/contants/constants";
-import { Gesture } from "../zoom/gesture";
-import { Zoom } from "../zoom/zoom";
+import { Colors } from "../../components/depth-chart/helpers";
+import { range } from "../../helpers";
+import { Renderer } from "../../renderer";
+import { Container } from "../../renderer/display";
+import { Graphics } from "../../renderer/graphics";
+import { InteractionEvent } from "../../renderer/interaction/interaction-event";
+import { Text } from "../../renderer/text";
+import { ScaleTime } from "../../types";
+import { AXIS_HEIGHT, FONT_SIZE } from "../../util/constants";
+import { Gesture, Zoom } from "../../util/zoom";
 import { pointer } from "./vertical-axis";
 
 const MAX_TICKS = 20;
