+++ src/ui/display-objects/line-curve.ts
@@ -1,11 +1,11 @@
-import { CurveFactory, curveStepBefore } from "d3-shape";
+import { CurveFactory, curveLinear } from "d3-shape";
 import { memoize, values } from "lodash";
 
-import { Texture } from "../../../renderer";
-import { Container } from "../../../renderer/display";
-import { Graphics } from "../../../renderer/graphics";
-import { Rectangle } from "../../../renderer/math";
-import { hex2string } from "../../../renderer/utils";
+import { Texture } from "../../renderer";
+import { Container } from "../../renderer/display";
+import { Graphics } from "../../renderer/graphics";
+import { Rectangle } from "../../renderer/math";
+import { hex2string } from "../../renderer/utils";
 
 function createGradTexture(
   colorStop1: number,
@@ -22,7 +22,7 @@ function createGradTexture(
 
   // adjust it if somehow you need better quality for very very big images
   canvas.width = 1;
-  canvas.height = Math.max(1, Math.abs(size + offset));
+  canvas.height = size + offset;
 
   const ctx = canvas.getContext("2d");
 
@@ -53,18 +53,8 @@ const createTexture = memoize(
       invert
     );
 
-    gradTexture.orig = new Rectangle(
-      0,
-      0,
-      1,
-      Math.max(1, Math.abs(size + offset))
-    );
-    gradTexture._frame = new Rectangle(
-      0,
-      0,
-      1,
-      Math.max(1, Math.abs(size + offset))
-    );
+    gradTexture.orig = new Rectangle(0, 0, 1, size + offset);
+    gradTexture._frame = new Rectangle(0, 0, 1, size + offset);
 
     return gradTexture;
   },
@@ -72,14 +62,12 @@ const createTexture = memoize(
 );
 
 /**
- * Draws an area curve
+ * Draws a line curve
  */
-export class PriceCurve extends Container {
-  private area: Graphics = new Graphics();
+export class LineCurve extends Container {
   private line: Graphics = new Graphics();
 
   private stroke: number;
-  private fill: number;
   private backgroundSurface: number;
   private curve: CurveFactory;
 
@@ -87,19 +75,16 @@ export class PriceCurve extends Container {
     stroke: number = 0,
     fill: number = 0xffffff,
     backgroundSurface: number = 0xffffff,
-    curve: CurveFactory = curveStepBefore
+    curve: CurveFactory = curveLinear
   ) {
     super();
 
     this.stroke = stroke;
-    this.fill = fill;
     this.backgroundSurface = backgroundSurface;
     this.curve = curve;
 
-    this.area.lineStyle({ width: 0 });
     this.line.lineStyle({ width: 4, color: stroke, alpha: 0.5 });
 
-    this.addChild(this.area);
     this.addChild(this.line);
   }
 
@@ -113,24 +98,8 @@ export class PriceCurve extends Container {
     invert: boolean,
     startPrice: number
   ): void {
-    this.fill = fill;
     this.stroke = stroke;
     this.backgroundSurface = backgroundSurface;
-    this.area.clear();
-
-    this.area.beginTextureFill({
-      alpha: 0.5,
-      texture: createTexture(
-        this.stroke,
-        this.backgroundSurface,
-        height,
-        startPrice,
-        invert
-      ),
-    });
-
-    this.area.drawArea(points, this.curve, startPrice);
-    this.area.endFill();
 
     this.line.clear();
     this.line.lineStyle({ width: 4, color: this.stroke });
