+++ src/ui/display-objects/rect.ts
@@ -1,4 +1,4 @@
-import { Graphics } from "../../../renderer/graphics";
+import { Graphics } from "../../renderer/graphics";
 
 /**
  * Draw a rectangle
